+++ packages/components/src/components/dialog/sando-dialog.stories.ts
@@ -1,0 +1,11 @@
+/**
+ * Sando Dialog Component Stories
+ *
+ * Stories will be added separately by sando-storybook.
+ */
+
+// Stories will be added separately
+export default {
+  title: 'Components/Dialog',
+  component: 'sando-dialog'
+};